+++ src/main/java/com/matsinger/barofishserver/utils/fcm/DummyFcmAdapter.java
@@ -4,13 +4,15 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 @RequiredArgsConstructor
 @Profile("local")
-public class DummyFcmService implements FcmService {
+public class DummyFcmAdapter implements FcmAdapter {
 
     @Override
-    public void sendFcmByToken(FcmRequestDto requestDto) {
+    public void sendMessage(List<FcmToken> tokens, FcmRequestDto requestDto) {
         System.out.println("📢 [Dummy FCM] " + requestDto.getTitle() + ": " + requestDto.getBody() + " (to: " + requestDto.getTargetUserId() + ")");
     }
 }
