+++ src/main/java/com/matsinger/barofishserver/utils/fcm/FcmAdapter.java
@@ -1,0 +1,8 @@
+package com.matsinger.barofishserver.utils.fcm;
+
+import java.util.List;
+
+public interface FcmAdapter {
+
+    void sendMessage(List<FcmToken> tokens, FcmRequestDto requestDto);
+}
