+++ src/main/java/com/matsinger/barofishserver/utils/fcm/FcmService.java
@@ -1,5 +1,18 @@
 package com.matsinger.barofishserver.utils.fcm;
 
-public interface FcmService {
-    void sendFcmByToken(FcmRequestDto requestDto);
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+public class FcmService {
+    private final FcmTokenRepository fcmTokenRepository;
+    private final FcmAdapter fcmAdapter;
+
+    public void sendFcmByToken(FcmRequestDto requestDto) {
+        List<FcmToken> tokens = fcmTokenRepository.findAllByUserId(requestDto.getTargetUserId());
+        fcmAdapter.sendMessage(tokens, requestDto);
+    }
 }
